MCKIN-8827: Cache updater data.
