+++ completion_aggregator/tasks/aggregation_tasks.py
@@ -5,30 +5,24 @@ Asynchronous tasks for performing aggregation of completions.
 from __future__ import absolute_import, division, print_function, unicode_literals
 
 import logging
-from collections import namedtuple
-from datetime import datetime
 
-import pytz
 from celery import shared_task
 from celery_utils.logged_task import LoggedTask
 from opaque_keys import InvalidKeyError
 from opaque_keys.edx.keys import CourseKey, UsageKey
-from xblock.completable import XBlockCompletionMode
-from xblock.core import XBlock
 
 from django.contrib.auth.models import User
 from django.db import connection
-from django.utils import timezone
 
-from .. import compat
-from ..models import Aggregator, StaleCompletion
-from ..utils import BagOfHolding
+from .. import core
+from ..models import StaleCompletion
 
 try:
     from progress.models import CourseModuleCompletion
     PROGRESS_IMPORTED = True
 except ImportError:
     PROGRESS_IMPORTED = False
+
 
 # SQLite doesn't support the ON DUPLICATE KEY syntax.  INSERT OR REPLACE will
 # have a similar effect, but uses new primary keys.  The drawbacks of this are:
@@ -55,10 +49,6 @@ INSERT_OR_UPDATE_SQLITE = """
 """
 
 
-OLD_DATETIME = pytz.utc.localize(datetime(1900, 1, 1, 0, 0, 0))
-
-CompletionStats = namedtuple('CompletionStats', ['earned', 'possible', 'last_modified'])
-
 log = logging.getLogger(__name__)
 
 
@@ -80,7 +70,6 @@ def update_aggregators(username, course_key, block_keys=(), force=False):
 
     Takes a collection of block_keys that have been updated, to enable future
     optimizations in how aggregators are recalculated.
-
     """
     try:
         user = User.objects.get(username=username)
@@ -92,227 +81,7 @@ def update_aggregators(username, course_key, block_keys=(), force=False):
     course_key = CourseKey.from_string(course_key)
     block_keys = set(UsageKey.from_string(key).map_into_course(course_key) for key in block_keys)
     log.info("Updating aggregators in %s for %s. Changed blocks: %s", course_key, user.username, block_keys)
-    return _update_aggregators(user, course_key, block_keys, force)
-
-
-def _update_aggregators(user, course_key, block_keys=frozenset(), force=False):
-    """
-    Update the aggregators for the specified enrollment (user + course).
-
-    This is the workhorse function for the update_aggregators task, taking its
-    arguments strongly typed.
-
-    Parameters
-    ----------
-        user (django.contrib.auth.models.User):
-            The user whose aggregators need updating.
-        course_key (opaque_keys.edx.keys.CourseKey):
-            The course in which the aggregators need updating.
-        block_keys (list[opaque_keys.edx.keys.UsageKey]):
-             A list of completable blocks that have changed.
-        force (bool):
-            If True, update aggregators even if they are up-to-date.
-
-    """
-    try:
-        updater = AggregationUpdater(user, course_key, compat.get_modulestore())
-    except compat.get_item_not_found_error():
-        log.exception("Course not found in modulestore.  Skipping aggregation for %s/%s.", user, course_key)
-    except TypeError:
-        log.exception("Could not parse modulestore data.  Skipping aggregation for %s/%s.", user, course_key)
-    else:
-        updater.update(block_keys, force)
-
-
-def calculate_updated_aggregators(user, course_key, changed_blocks=frozenset(), root_block=None, force=False):
-    try:
-        updater = AggregationUpdater(user, course_key, compat.get_modulestore(), root_block=root_block)
-    except compat.get_item_not_found_error():
-        log.exception("Course not found in modulestore.  Skipping aggregation for %s/%s.", user, course_key)
-        return []
-    except TypeError:
-        log.exception("Could not parse modulestore data.  Skipping aggregation for %s/%s.", user, course_key)
-        return []
-    else:
-        return updater.calculate_updated_aggregators(changed_blocks, force)
-
-
-class AggregationUpdater(object):
-    """
-    Class to update aggregators for a given course and user.
-    """
-
-    def __init__(self, user, course_key, modulestore, root_block=None):
-        """
-        Create an aggregation updater for the given user and course.
-
-        Also takes a modulestore instance.
-        """
-        self.user = user
-        self.course_key = course_key
-
-        with modulestore.bulk_operations(self.course_key):
-            if root_block:
-                self.root_block = root_block
-            else:
-                self.root_block = compat.init_course_block_key(modulestore, self.course_key)
-            self.course_blocks = compat.init_course_blocks(self.user, self.root_block)
-
-        self.aggregators = {
-            aggregator.block_key: aggregator for aggregator in Aggregator.objects.filter(
-                user=self.user,
-                course_key=self.course_key,
-            )
-        }
-        # used to store all rows for update
-        self.updated_aggregators = []
-        self.block_completions = {
-            completion.block_key.map_into_course(self.course_key): completion
-            for completion in compat.get_block_completions(self.user, self.course_key)
-        }
-
-    def get_affected_aggregators(self, changed_blocks):
-        """
-        Return the set of aggregator blocks that may need updating.
-        """
-        if changed_blocks:
-            return compat.get_affected_aggregators(self.course_blocks, changed_blocks)
-        else:
-            return BagOfHolding()
-
-    def calculate_updated_aggregators(self, changed_blocks=frozenset(), force=False):
-        """
-        Return updated aggregators without sumitting them to the database.
-
-        And without clearing stale completions.
-        """
-        affected_aggregators = self.get_affected_aggregators(changed_blocks)
-        self.update_for_block(self.root_block, affected_aggregators, force)
-        return self.updated_aggregators
-
-    def update(self, changed_blocks=frozenset(), force=False):
-        """
-        Update the aggregators for the course.
-
-        Takes a set of completable blocks that have been recently
-        updated to inform how to perform the update. If supplied, only
-        the aggregators containing those blocks will be
-        updated. Otherwise, the entire course tree will be updated.
-        """
-        start = timezone.now()
-        updated_aggregators = self.calculate_updated_aggregators(changed_blocks, force)
-        Aggregator.objects.bulk_create_or_update(updated_aggregators)
-        self.resolve_stale_completions(changed_blocks, start)
-
-    def update_for_block(self, block, affected_aggregators, force=False):
-        """
-        Recursive function to perform updates for a given block.
-
-        Dispatches to an appropriate method given the block's completion_mode.
-        """
-        mode = XBlockCompletionMode.get_mode(XBlock.load_class(block.block_type))
-        if mode == XBlockCompletionMode.EXCLUDED:
-            return self.update_for_excluded()
-        elif mode == XBlockCompletionMode.COMPLETABLE:
-            return self.update_for_completable(block)
-        elif mode == XBlockCompletionMode.AGGREGATOR:
-            return self.update_for_aggregator(block, affected_aggregators, force)
-        else:
-            raise ValueError("Invalid completion mode {}".format(mode))
-
-    def update_for_aggregator(self, block, affected_aggregators, force):
-        """
-        Calculate the new completion values for an aggregator.
-        """
-        total_earned = 0.0
-        total_possible = 0.0
-        last_modified = OLD_DATETIME
-
-        if block not in affected_aggregators:
-            try:
-                obj = Aggregator.objects.get(user=self.user, course_key=self.course_key, block_key=block)
-            except Aggregator.DoesNotExist:
-                pass
-            else:
-                return CompletionStats(earned=obj.earned, possible=obj.possible, last_modified=obj.last_modified)
-        for child in compat.get_children(self.course_blocks, block):
-            (earned, possible, modified) = self.update_for_block(child, affected_aggregators, force)
-            total_earned += earned
-            total_possible += possible
-            if modified is not None:
-                last_modified = max(last_modified, modified)
-        if self._aggregator_needs_update(block, last_modified, force):
-            if total_possible == 0.0:
-                percent = 1.0
-            else:
-                percent = total_earned / total_possible
-            Aggregator.objects.validate(self.user, self.course_key, block)
-            if block not in self.aggregators:
-                aggregator = Aggregator(
-                    user=self.user,
-                    course_key=self.course_key,
-                    block_key=block,
-                    aggregation_name=block.block_type,
-                    earned=total_earned,
-                    possible=total_possible,
-                    percent=percent,
-                    last_modified=last_modified,
-                )
-                self.aggregators[block] = aggregator
-            else:
-                aggregator = self.aggregators[block]
-                aggregator.earned = total_earned
-                aggregator.possible = total_possible
-                aggregator.percent = percent
-                aggregator.last_modified = last_modified
-                aggregator.modified = timezone.now()
-            self.updated_aggregators.append(aggregator)
-        return CompletionStats(earned=total_earned, possible=total_possible, last_modified=last_modified)
-
-    def update_for_excluded(self):
-        """
-        Return a sentinel empty completion value for excluded blocks.
-        """
-        return CompletionStats(earned=0.0, possible=0.0, last_modified=OLD_DATETIME)
-
-    def update_for_completable(self, block):
-        """
-        Return the block completion value for a given completable block.
-        """
-        completion = self.block_completions.get(block)
-        if completion:
-            earned = completion.completion
-            last_modified = completion.modified
-        else:
-            earned = 0.0
-            last_modified = OLD_DATETIME
-        return CompletionStats(earned=earned, possible=1.0, last_modified=last_modified)
-
-    def _aggregator_needs_update(self, block, modified, force):
-        """
-        Return True if the given aggregator block needs to be updated.
-
-        This method assumes that the block has already been determined to be an aggregator.
-        """
-        if Aggregator.block_is_registered_aggregator(block):
-            agg = self.aggregators.get(block)
-            if agg is None or force:
-                return True
-            return getattr(agg, 'last_modified', OLD_DATETIME) < modified
-        return False
-
-    def resolve_stale_completions(self, changed_blocks, start):
-        """
-        Find all StaleCompletions resolved by this task and mark them resolved.
-        """
-        queryset = StaleCompletion.objects.filter(
-            username=self.user.username,
-            course_key=self.course_key,
-            modified__lt=start,
-        )
-        if changed_blocks:
-            queryset = queryset.filter(block_key__in=changed_blocks)
-        queryset.update(resolved=True)
+    return core.update_aggregators(user, course_key, block_keys, force)
 
 
 @shared_task
